requirements.txt
absl-py==0.9.0
astor==0.8.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cycler==0.10.0
Cython==0.29.17
gast==0.2.2
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
Markdown==3.2.1
matplotlib==3.2.1
MouseInfo==0.1.3
mss==5.0.0
numpy==1.18.3
oauthlib==3.1.0
object-detection==0.1
opencv-python==4.2.0.34
opt-einsum==3.2.1
Pillow==7.1.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.50
PyGetWindow==0.0.8
PyMsgBox==1.0.7
pyparsing==2.4.7
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
PyTweening==1.0.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scipy==1.4.1
six==1.14.0
slim==0.1
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
tensorflow-gpu-estimator==2.1.0
termcolor==1.1.0
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
